KB-26: Moved everything in `IncubatorContents` to `IncubatorPage`.
